feat: ジョブのステータスを「completed」から「succeeded」に変更し、Observability設定を追加
- job_handler.rs: ジョブのステータスを「completed」から「succeeded」に変更
- scheduler_execution_postgres.rs: ステータスに「triggered_by」を追加し、SQLクエリを更新
- config.rs: ObservabilityConfigを追加し、設定項目を定義
- cron_engine.rs: ジョブの実行完了時にステータスを「succeeded」に更新
- main.rs: Telemetry設定をObservabilityから取得するように変更
- trigger_job.rs: ジョブ実行完了時のステータスを「succeeded」に変更
- config.yaml: Observability設定を追加
- search/config.rs: ObservabilityConfigを追加
- search/main.rs: Telemetry設定をObservabilityから取得するように変更
- session/config.rs: ObservabilityConfigを追加
- session/main.rs: Telemetry設定をObservabilityから取得するように変更
- tenant/config.yaml: Observability設定を追加
- tenant/main.rs: Telemetry設定をObservabilityから取得するように変更
- vault/config.yaml: Observability設定を追加
- vault/main.rs: Telemetry設定をObservabilityから取得するように変更
- workflow/config.yaml: Observability設定を追加
- workflow/main.rs: Telemetry設定をObservabilityから取得するように変更

+++ regions/system/server/rust/quota/src/adapter/repository/quota_policy_postgres.rs
@@ -26,7 +26,7 @@ struct QuotaPolicyRow {
     quota_limit: i64,
     period: String,
     enabled: bool,
-    alert_threshold_percent: f64,
+    alert_threshold_percent: i16,
     created_at: DateTime<Utc>,
     updated_at: DateTime<Utc>,
 }
@@ -113,7 +113,7 @@ impl QuotaPolicyRepository for QuotaPolicyPostgresRepository {
         .bind(policy.limit as i64)
         .bind(policy.period.as_str())
         .bind(policy.enabled)
-        .bind(policy.alert_threshold_percent.unwrap_or(80) as f64)
+        .bind(policy.alert_threshold_percent.unwrap_or(80) as i16)
         .bind(policy.created_at)
         .bind(policy.updated_at)
         .execute(self.pool.as_ref())
@@ -139,7 +139,7 @@ impl QuotaPolicyRepository for QuotaPolicyPostgresRepository {
         .bind(policy.limit as i64)
         .bind(policy.period.as_str())
         .bind(policy.enabled)
-        .bind(policy.alert_threshold_percent.unwrap_or(80) as f64)
+        .bind(policy.alert_threshold_percent.unwrap_or(80) as i16)
         .bind(policy.updated_at)
         .execute(self.pool.as_ref())
         .await?;
